Fix 👨‍🔧: RPN no longer tries to devide by 0, and values are now doubles to not lose precision in devisions

+++ ex01/includes/RPN.h
@@ -5,7 +5,7 @@
 #include <stack>
 
 
-int		calculate(std::string input);
+double		calculate(std::string input);
 void	parse(std::string & parsed, std::string raw);
